requirements.txt
alembic==1.6.5
annotated-types==0.7.0
astroid==2.5.6
blinker==1.8.2
boto3==1.34.131
botocore==1.34.132
cachetools==5.5.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
geographiclib==2.0
geopy==2.4.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-vision==3.4.5
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.0
grpcio==1.66.0
grpcio-status==1.62.3
gunicorn==20.1.0
httplib2==0.22.0
idna==3.7
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==1.0.1
lazy-object-proxy==1.6.0
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
pillow==10.3.0
pillow-heif==0.18.0
proto-plus==1.24.0
protobuf==4.25.4
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pylint==2.8.2
pyparsing==3.1.2
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-http-client==3.3.7
requests==2.32.3
requests-aws4auth==1.2.3
rsa==4.9
s3transfer==0.10.1
sendgrid==6.11.0
setuptools==73.0.1
six==1.16.0
SQLAlchemy==1.4.11
starkbank-ecdsa==2.2.0
toml==0.10.2
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.19
Werkzeug==1.0.1
wrapt==1.12.1
reportlab==4.0.4
qrcode==7.4.2
Pillow==10.0.0
